Rename a few variables and functions from base to urlGen
