+++ __TESTS_BUNDLE_SIZE__/utils/stringGenerators/importFromPackage.ts
@@ -1,10 +1,10 @@
 /**
- * Utility function used to import items from the index of Base
+ * Utility function used to import items from the index of the project
  * @param {string} exportedObject - The name of an exported object in index, for example ImageTransformation.
  *                                  The result will be `import {ImageTransformation} from '../../dist'
  * @returns string
  */
-function importFromBase(exportedObject: string): string {
+function importFromPackage(exportedObject: string): string {
   return `
     import {${exportedObject}} from "${process.cwd()}/dist";
     // we console log to force the bundle not to tree shake
@@ -12,4 +12,4 @@ function importFromBase(exportedObject: string): string {
   `;
 }
 
-export default importFromBase;
+export default importFromPackage;
