--- src/actions.ts
@@ -1,41 +1,0 @@
-/**
- * @namespace Actions
- */
-
-import {RoundCorners} from "./actions/roundCorners";
-import {Resize} from "./actions/resize";
-import {Border} from "./actions/border";
-import {Effect} from "./actions/effect";
-import {Rotate} from "./actions/rotate";
-import {Adjust} from "./actions/adjust";
-import {VideoEdit} from "./actions/videoEdit";
-import {Overlay} from "./actions/overlay";
-import {Underlay} from "./actions/underlay";
-import {NamedTransformation} from './actions/namedTransformation';
-import {Delivery} from "./actions/delivery";
-import {CustomFunction} from './actions/customFunction';
-import {Transcode} from "./actions/transcode";
-import {PSDTools} from "./actions/psdTools";
-import {Animated} from "./actions/animated";
-
-const Actions = {
-  Resize,
-  Border,
-  RoundCorners,
-  Effect,
-  Rotate,
-  Adjust,
-  Overlay,
-  Underlay,
-  NamedTransformation,
-  Delivery,
-  CustomFunction,
-  VideoEdit,
-  Transcode,
-  PSDTools,
-  Animated
-};
-
-export {
-  Actions
-};
